backend/requirements.txt
newspaper3k==0.2.8
openai==0.28.0
python-dotenv==1.0.1
lxml==5.2.1
lxml_html_clean==0.1.1
